New application to be used for db tests to populate oracle DB. I have split testEcalObjects into a main and a class with several methods. the code is better written now and is more general.
It requires command line arguments to specify the connection string, that can be oracle sqlite or anything, the tag
to be aplied to the objects and also the first and last runs to be used to store objects into DB.

for now the pedestals are commented out in order to prevent intereference with data stored already by Rick.

Soon we'll be using just this one application to populate any type of source.

It will also replace testEcalObject in future releases